+++ modules/sensor-extraction/ros-to-png/tests/test_stack.py
@@ -11,9 +11,9 @@ from aws_cdk.assertions import Template
 
 @pytest.fixture(scope="function")
 def stack_defaults():
-    os.environ["ADDF_PROJECT_NAME"] = "test-project"
-    os.environ["ADDF_DEPLOYMENT_NAME"] = "test-deployment"
-    os.environ["ADDF_MODULE_NAME"] = "test-module"
+    os.environ["SEEDFARMER_PROJECT_NAME"] = "test-project"
+    os.environ["SEEDFARMER_DEPLOYMENT_NAME"] = "test-deployment"
+    os.environ["SEEDFARMER_MODULE_NAME"] = "test-module"
     os.environ["CDK_DEFAULT_ACCOUNT"] = "111111111111"
     os.environ["CDK_DEFAULT_REGION"] = "us-east-1"
 
@@ -25,6 +25,7 @@ def test_synthesize_stack(stack_defaults):
     import stack
 
     app = cdk.App()
+    project_name = "test-project"
     dep_name = "test-deployment"
     mod_name = "test-module"
 
@@ -39,9 +40,11 @@ def test_synthesize_stack(stack_defaults):
 
     ros_to_png = stack.RosToPngBatchJob(
         scope=app,
-        id=f"addf-{dep_name}-{mod_name}",
+        id=f"{project_name}-{dep_name}-{mod_name}",
+        project_name=project_name,
         deployment_name=dep_name,
         module_name=mod_name,
+        ecr_repository_arn="arn:aws:ecr:us-east-1:123456789012:repository/addf-docker-repository",
         s3_access_policy="arn:aws:iam::123456789012:policy/addf-buckets-us-west-2-123-full-access",
         batch_config=batch_config,
         stack_description="Testing",
@@ -52,17 +55,9 @@ def test_synthesize_stack(stack_defaults):
     )
 
     template = Template.from_stack(ros_to_png)
-    template.resource_count_is("AWS::ECR::Repository", 1)
-    template.resource_count_is("AWS::Lambda::Function", 2)
+    template.resource_count_is("AWS::Lambda::Function", 1)
     template.resource_count_is("AWS::Batch::JobDefinition", 1)
-    template.resource_count_is("AWS::IAM::Role", 3)
-    # Check ecr.Repository 'auto_delete' runtime version
-    template.has_resource_properties(
-        type="AWS::Lambda::Function",
-        props={
-            "Runtime": "nodejs18.x",
-        },
-    )
+    template.resource_count_is("AWS::IAM::Role", 2)
     # Check batch job definition properties
     template.has_resource_properties(
         type="AWS::Batch::JobDefinition",
@@ -90,6 +85,7 @@ def test_synthesize_stack_without_resize(stack_defaults):
     import stack
 
     app = cdk.App()
+    project_name = "test-project"
     dep_name = "test-deployment"
     mod_name = "test-module"
 
@@ -102,9 +98,11 @@ def test_synthesize_stack_without_resize(stack_defaults):
 
     ros_to_png = stack.RosToPngBatchJob(
         scope=app,
-        id=f"addf-{dep_name}-{mod_name}",
+        id=f"{project_name}-{dep_name}-{mod_name}",
+        project_name=project_name,
         deployment_name=dep_name,
         module_name=mod_name,
+        ecr_repository_arn="arn:aws:ecr:us-east-1:123456789012:repository/addf-docker-repository",
         s3_access_policy="arn:aws:iam::123456789012:policy/addf-buckets-us-west-2-123-full-access",
         batch_config=batch_config,
         stack_description="Testing",
@@ -115,7 +113,6 @@ def test_synthesize_stack_without_resize(stack_defaults):
     )
 
     template = Template.from_stack(ros_to_png)
-    template.resource_count_is("AWS::ECR::Repository", 1)
-    template.resource_count_is("AWS::Lambda::Function", 2)
+    template.resource_count_is("AWS::Lambda::Function", 1)
     template.resource_count_is("AWS::Batch::JobDefinition", 1)
-    template.resource_count_is("AWS::IAM::Role", 3)
+    template.resource_count_is("AWS::IAM::Role", 2)